Support read-only requests
Readonly requests allow safely reading from a statemachine without
requiring an entry be appended to be log. This may be more performant in
certain situations. A few performance notes:
- This requires the same amount of communication as `propose` (a
majority of nodes need to consent on a valid read index)
- This doesn't require appending an entry to the log, so it reduces
snapshot pressure (when implemented) and any persistence costs (if/when
implemented).

If `local=True` is provided to `read`, a local read is used. This
provides serializable guarantees, but not linearizable guarantees.

The demo has been updated to support a query parameter `read` which can
take either `global` (for linearizable reads), `local` (for serializable
reads), or `propose` (for using propose to do a read).